Feat: HCI gotov

+++ booking/booking/WPF/ViewModels/TourCancellationViewModel.cs
@@ -13,6 +13,10 @@ using booking.Repository;
 using booking.WPF.ViewModels;
 using Microsoft.Web.WebView2.Core.Raw;
 using System.ComponentModel;
+using System.Threading;
+using Domain.DTO;
+using System.Threading.Tasks;
+using WPF.Views.Guide;
 
 namespace WPF.ViewModels
 {
@@ -30,6 +34,31 @@ namespace WPF.ViewModels
                 {
                     selectedTour = value;
                     OnPropertyChanged(nameof(SelectedTour));
+                }
+            }
+        }
+        private CancellationTokenSource cts = new CancellationTokenSource();
+
+        private bool demoOn;
+        public bool DemoOn
+        {
+            get { return demoOn; }
+            set
+            {
+                demoOn = value;
+                OnPropertyChanged(nameof(DemoOn));
+            }
+        }
+        private string demoName;
+        public string DemoName
+        {
+            get { return demoName; }
+            set
+            {
+                if (demoName != value)
+                {
+                    demoName = value;
+                    OnPropertyChanged(nameof(DemoName));
                 }
             }
         }
@@ -58,10 +87,11 @@ namespace WPF.ViewModels
             UpcomingTours = _tourService.FindUpcomingTours();
             //SelectedTour = new Tour();
             Guide = guide;
+            DemoName = "Demo";
         }
         
         public ICommand ExitCommand => new RelayCommand(ExitWindow);
-
+        public ICommand DemoCommand => new RelayCommand(StartStopDemo);
         public ICommand TooltipToursCommand => new RelayCommand(ToolTipTourShow);
         public void ToolTipTourShow()
         {
@@ -82,11 +112,14 @@ namespace WPF.ViewModels
         {
             if (SelectedTour.Id > 0)
             {
-                if (MessageBox.Show("Are you sure?", "Warning", MessageBoxButton.YesNo,MessageBoxImage.Warning) == MessageBoxResult.Yes)
+                if (MessageBox.Show("Are you sure?", "Warning", MessageBoxButton.YesNo, MessageBoxImage.Warning) == MessageBoxResult.Yes)
                 {
-                    _reservationService.GiveVouchers(SelectedTour,Guide);
-                    _tourService.DeleteTour(SelectedTour);
-                    UpcomingTours.Remove(SelectedTour);
+                    if (!DemoOn)
+                    { 
+                        _reservationService.GiveVouchers(SelectedTour, Guide);
+                        _tourService.DeleteTour(SelectedTour);
+                        UpcomingTours.Remove(SelectedTour);
+                    }   
                     //ExitWindow();
                 }
             }
@@ -98,6 +131,51 @@ namespace WPF.ViewModels
         private void ExitWindow()
         {
             this.CloseCurrentWindow();
+        }
+        private void StartStopDemo()
+        {
+            if (DemoOn)
+            {
+                cts.Cancel();
+                DemoOn = !DemoOn;
+                DemoName = "Demo";
+                MessageBox.Show("Demo has been stopped!", "Demo message", MessageBoxButton.OK, MessageBoxImage.Warning);
+            }
+            else
+            {
+                cts = new CancellationTokenSource();
+                DemoOn = !DemoOn;
+                DemoIsOn(cts.Token);
+                DemoName = "Stop";
+            }
+        }
+        private async Task DemoIsOn(CancellationToken ct)
+        {
+
+            if (DemoOn)
+            {
+
+                ct.ThrowIfCancellationRequested();
+                StartDemoWindow sdw = new StartDemoWindow();
+                sdw.Show();
+
+                await Task.Delay(2000, ct);
+                sdw.Close();
+                SelectedTour = UpcomingTours[0];
+
+                await Task.Delay(2000, ct);
+                AbandonTour();
+                
+                await Task.Delay(1000, ct);
+                FinishedDemoWindow fdw = new FinishedDemoWindow();
+                fdw.Show();
+
+                await Task.Delay(2000, ct);
+                fdw.Close();
+                DemoOn = !DemoOn;
+                DemoName = "Demo";
+
+            }
         }
     }
 }
